requirements.txt
Django==4.0.3
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
psycopg2-binary==2.9.3
django-mptt
django-seed==0.3.1